Fixes companies verification button
